- Added ProgressBar
- Changed game loading percentage to 0-1 value instead of 0-100
- Added Lime Color
- Added FillStyle for ProgressBar
- Added updating anchoring system for all GuiItems
- Removed accessibility of text mesh in Label
- Shader will now log the source if a attribute or uniform cannot be found
- Added tiny delay to loading tick to normal tick
- Changed way borderless fullscreen worked in Window

+++ code/include/GameLogic/DynamicObject.h
@@ -49,7 +49,7 @@ namespace Plutonium
 
 		Game *parent;
 		DynamicModel *model;
-		int percentage;
+		float percentage;
 		Initializer initializer;
 
 		bool running;
